Compartmentalize components
