+++ flat_out_app/lib/core/ui_functional_components/flow_notifier.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart';
+
+class FlowNotifier extends ChangeNotifier {
+  
+}
