+++ flat_out_app/lib/screens/auth_flow/auth_flow_user_components/header/header_animated_slide.dart
@@ -1,0 +1,57 @@
+import 'package:flat_out_app/core/ui_functional_components/ui_statics.dart';
+import 'package:flutter/cupertino.dart';
+
+class HeaderAnimatedSlide extends StatefulWidget {
+  final AnimationController controller;
+
+
+  HeaderAnimatedSlide({required this.controller});
+
+  @override
+  State<StatefulWidget> createState() => _HeaderAnimatedSlideState();
+}
+
+class _HeaderAnimatedSlideState extends State<HeaderAnimatedSlide> {
+  late final Tween<double> _pointTween =
+  Tween<double>(begin: UiStatics.headerInset, end: MediaQuery.of(context).size.width - UiStatics.headerInset);
+  late final Animation<double> _pointAnimation = _pointTween.animate(widget.controller)
+    ..addListener(() {
+      setState(() {});
+    });
+
+  @override
+  Widget build(BuildContext context) {
+    return CustomPaint(
+      painter: ShapePainter(value: _pointAnimation.value),
+      size: Size(MediaQuery.of(context).size.width, 50),
+    );
+  }
+}
+
+class ShapePainter extends CustomPainter {
+  double value;
+
+  ShapePainter({required this.value}) : super();
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    Paint paint = Paint()..color = Color(0xff383e64);
+
+    Path path = Path();
+    path.moveTo(0, -1);
+    path.lineTo(size.width, -1);
+    path.lineTo(size.width, size.height);
+    // path.quadraticBezierTo(value, -20, 0, size.height);
+    path.lineTo(value, size.height / 3);
+    path.lineTo(0, size.height);
+
+    // canvas.drawShadow(path, Colors.black, 5, true);
+    canvas.drawPath(path, paint);
+  }
+
+  @override
+  bool shouldRepaint(CustomPainter oldDelegate) {
+    return true;
+  }
+}
+
