--- flat_out_app/lib/screens/auth_flow/auth_flow_user_components/user_auth_header.dart
@@ -1,129 +1,0 @@
-import 'package:flat_out_app/core/ui_functional_components//ui_statics.dart';
-import 'package:flutter/material.dart';
-
-class UserAuthHeader extends StatefulWidget {
-  late final bool isLoginPage;
-  late final void Function(bool) swapPage;
-  
-  UserAuthHeader({required this.isLoginPage, required this.swapPage});
-
-  @override
-  State<StatefulWidget> createState() => _UserAuthHeaderState();
-}
-
-class _UserAuthHeaderState extends State<UserAuthHeader> with TickerProviderStateMixin {
-  
-  late final AnimationController _controller = AnimationController(vsync: this, duration: Duration(milliseconds: 300));
-  late final Tween<double> _pointTween =
-      Tween<double>(begin: UiStatics.headerInset, end: MediaQuery.of(context).size.width - UiStatics.headerInset);
-  late final Animation<double> _pointAnimation = _pointTween.animate(_controller)
-    ..addListener(() {
-      setState(() {});
-    });
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-
-    return Container(
-      child: Column(
-        children: [
-          Container(
-            width: MediaQuery.of(context).size.width,
-            padding: EdgeInsets.only(top: MediaQuery.of(context).size.height / 20),
-            color: Theme.of(context).dividerColor,
-            child: Column(
-              children: [
-                Icon(
-                  Icons.real_estate_agent_outlined,
-                  color: Theme.of(context).scaffoldBackgroundColor,
-                  size: MediaQuery.of(context).size.height / 10,
-                ),
-                Text(
-                  "Flat-Out Management",
-                  textAlign: TextAlign.center,
-                  style: Theme.of(context).textTheme.headline3,
-                ),
-              ],
-            ),
-          ),
-          CustomPaint(
-            painter: ShapePainter(value: _pointAnimation.value),
-            size: Size(MediaQuery.of(context).size.width, 50),
-          ),
-          Container(
-            child: Row(
-              mainAxisAlignment: MainAxisAlignment.center,
-              children: [
-                TextButton(
-                    onPressed: () {
-                      FocusManager.instance.primaryFocus?.unfocus();
-                      widget.swapPage(true);
-                      _controller.reverse();
-                    },
-                    style: ButtonStyle(
-                        foregroundColor: widget.isLoginPage
-                            ? MaterialStateProperty.all(Theme.of(context).textTheme.headline4?.color)
-                            : MaterialStateProperty.all(Theme.of(context).textTheme.headline5?.color),
-                        textStyle: widget.isLoginPage
-                            ? MaterialStateProperty.all(Theme.of(context).textTheme.headline4)
-                            : MaterialStateProperty.all(Theme.of(context).textTheme.headline5)),
-                    child: Text("Login")),
-                Text(
-                  "|",
-                  style: Theme.of(context).textTheme.headline5,
-                ),
-                TextButton(
-                    onPressed: () {
-                      FocusManager.instance.primaryFocus?.unfocus();
-                      widget.swapPage(false);
-                      _controller.forward();
-                    },
-                    style: ButtonStyle(
-                        foregroundColor: !widget.isLoginPage
-                            ? MaterialStateProperty.all(Theme.of(context).textTheme.headline4?.color)
-                            : MaterialStateProperty.all(Theme.of(context).textTheme.headline5?.color),
-                        textStyle: !widget.isLoginPage
-                            ? MaterialStateProperty.all(Theme.of(context).textTheme.headline4)
-                            : MaterialStateProperty.all(Theme.of(context).textTheme.headline5)),
-                    child: Text("Signup"))
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class ShapePainter extends CustomPainter {
-  double value;
-
-  ShapePainter({required this.value}) : super();
-
-  @override
-  void paint(Canvas canvas, Size size) {
-    Paint paint = Paint()..color = Color(0xff383e64);
-
-    Path path = Path();
-    path.moveTo(0, -1);
-    path.lineTo(size.width, -1);
-    path.lineTo(size.width, size.height);
-    // path.quadraticBezierTo(value, -20, 0, size.height);
-    path.lineTo(value, size.height / 3);
-    path.lineTo(0, size.height);
-
-    // canvas.drawShadow(path, Colors.black, 5, true);
-    canvas.drawPath(path, paint);
-  }
-
-  @override
-  bool shouldRepaint(CustomPainter oldDelegate) {
-    return true;
-  }
-}
